refactor: moved the accessors
